Move all DIF data classes to separate package
