+++ vck-aries/src/commonMain/kotlin/at/asitplus/wallet/lib/msg/RequestCredentialAttachment.kt
@@ -2,8 +2,8 @@ package at.asitplus.wallet.lib.msg
 
 import at.asitplus.KmmResult.Companion.wrap
 import at.asitplus.wallet.lib.aries.jsonSerializer
-import at.asitplus.wallet.lib.data.dif.CredentialManifest
-import at.asitplus.wallet.lib.data.dif.PresentationSubmission
+import at.asitplus.dif.CredentialManifest
+import at.asitplus.dif.PresentationSubmission
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.encodeToString
