+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/helper/AuthorizationResponsePreparationState.kt
@@ -1,6 +1,6 @@
 package at.asitplus.wallet.lib.oidc.helpers
 
-import at.asitplus.wallet.lib.data.dif.PresentationDefinition
+import at.asitplus.dif.PresentationDefinition
 import at.asitplus.openid.RelyingPartyMetadata
 import kotlinx.serialization.Serializable
