+++ vck/src/commonMain/kotlin/at/asitplus/dif/ConstraintHolder.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.lib.data.dif
+package at.asitplus.dif
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -8,8 +8,9 @@ import kotlinx.serialization.Serializable
  * [DIF Presentation Exchange v1.0.0](https://identity.foundation/presentation-exchange/spec/v1.0.0/#presentation-definition)
  */
 @Serializable
-data class FormatContainerJwt(
-    // TODO make this a collection of Json Web Algorithms from kmp-crypto
-    @SerialName("alg")
-    val algorithms: Collection<String>? = null,
+data class ConstraintHolder(
+    @SerialName("field_id")
+    val fieldIds: Collection<String>,
+    @SerialName("directive")
+    val directive: RequirementEnum,
 )
