+++ vck/src/commonMain/kotlin/at/asitplus/dif/ConstraintStatus.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.lib.data.dif
+package at.asitplus.dif
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -8,9 +8,7 @@ import kotlinx.serialization.Serializable
  * [DIF Presentation Exchange v1.0.0](https://identity.foundation/presentation-exchange/spec/v1.0.0/#presentation-definition)
  */
 @Serializable
-data class ConstraintNotFilter(
-    @SerialName("const")
-    val const: String? = null,
-    @SerialName("enum")
-    val enum: Collection<String>? = null,
+data class ConstraintStatus(
+    @SerialName("directive")
+    val directive: StatusDirectiveEnum
 )
