+++ vck/src/commonMain/kotlin/at/asitplus/dif/ConstraintStatusHolder.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.lib.data.dif
+package at.asitplus.dif
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
