+++ vck/src/commonMain/kotlin/at/asitplus/dif/Json.kt
@@ -1,0 +1,12 @@
+package at.asitplus.dif
+
+import kotlinx.serialization.json.Json
+
+val jsonSerializer by lazy {
+    Json {
+        prettyPrint = false
+        encodeDefaults = false
+        classDiscriminator = "type"
+        ignoreUnknownKeys = true
+    }
+}
