+++ vck/src/commonMain/kotlin/at/asitplus/dif/PresentationDefinition.kt
@@ -1,7 +1,6 @@
-package at.asitplus.wallet.lib.data.dif
+package at.asitplus.dif
 
 import at.asitplus.KmmResult.Companion.wrap
-import at.asitplus.wallet.lib.data.vckJsonSerializer
 import com.benasher44.uuid.uuid4
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -11,7 +10,6 @@ import kotlinx.serialization.encodeToString
  * Data class for
  * [DIF Presentation Exchange v1.0.0](https://identity.foundation/presentation-exchange/spec/v1.0.0/#presentation-definition)
  */
-// TODO into data classes
 @Serializable
 data class PresentationDefinition(
     @SerialName("id")
@@ -32,11 +30,11 @@ data class PresentationDefinition(
         formats: FormatHolder
     ) : this(id = uuid4().toString(), inputDescriptors = inputDescriptors, formats = formats)
 
-    fun serialize() = vckJsonSerializer.encodeToString(this)
+    fun serialize() = jsonSerializer.encodeToString(this)
 
     companion object {
         fun deserialize(it: String) = kotlin.runCatching {
-            vckJsonSerializer.decodeFromString<PresentationDefinition>(it)
+            jsonSerializer.decodeFromString<PresentationDefinition>(it)
         }.wrap()
     }
 }
