+++ vck/src/commonMain/kotlin/at/asitplus/dif/StatusDirectiveEnumSerializer.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.lib.data.dif
+package at.asitplus.dif
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.PrimitiveKind
